Fix data type for weight creation

+++ src/main/java/com/boyarsky/apiservice/dto/measurement/CreateWeightDto.java
@@ -11,5 +11,5 @@ public class CreateWeightDto extends CreateMeasurementDto {
     @NotNull
     @Min(10)
     @Max(999)
-    private Integer weight;
+    private Float weight;
 }
